spi_flash: do not return error code for unsupported flash lock/unlock ops
Instead, print a warning message for the user to see.

This is, in order to support non-Micron/STMicro flashes for
internal/experimental usage, or for other boards that are similar to
Pluto/M2k, but with different flash types.

The flashing operation will try to do a `sf protect lock/unlock` on the
flash, and that will error out for non-Micron/STMicro flashes.
That's because other flash types don't support the ops [in the driver].

Disabling an error condition is not nice, but it's easier than modifying
the u-boot vars for each flash-type. And we still get a warning message, so
there should be awareness about this.

Supporting lock/unlock for ISSI flashes is possible, but requires more work
because ISSI needs to access another register for doing the same
lock/unlock as Micron/STmicro (i.e. access the Top/Bottom select & BP3
bit).

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>