sidekiqz2: update u-boot env settings with manufacturer specs
This is largely a copy of the extra envs from  `zynq-common.h`.
What's been adapted is that GPIO 48 controls the DFU boot mode.

These env vars have been copied/adapted from a running SidekiqZ2.

Maybe at some point [later] it could make sense to adapt these to be a bit
more granular [i.e. define macros for QSPIBOOT and so on], but for now they
should be fine.

Signed-off-by: Alexandru Ardelean <alexandru.ardelean@analog.com>